move {react,server}/index up and export wildcards
